Refactor: Split monolithic services.py into a services package
The single `services.py` file, which contained all service classes,
has been refactored into a `services` package. Each service class
now resides in its own file within this package.

An `__init__.py` file is used to maintain a consistent import path
for other parts of the application.

This change improves modularity and maintainability.
