+++ server/domain/services/__init__.py
@@ -1,0 +1,21 @@
+from .user_service import user_service
+from .site_service import site_service
+from .crane_service import crane_service
+from .assignment_service import assignment_service
+from .document_service import document_service
+from .attendance_service import attendance_service
+from .request_service import request_service
+from .owner_service import owner_service
+from .crane_model_service import crane_model_service
+
+__all__ = [
+    "user_service",
+    "site_service",
+    "crane_service",
+    "assignment_service",
+    "document_service",
+    "attendance_service",
+    "request_service",
+    "owner_service",
+    "crane_model_service",
+]
